准备stage2 更新完成
第一行加了label，已经晚上波形显示

+++ SPDisplay/DataCell.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 namespace SPDisplay {
     public class DataCell {
         public int line;
-        public ulong address;
-        public ulong datalength;
+        public int address;
+        public int datalength;
         public byte[] data;
     }
 }
